+++ docs/changes.md
@@ -1,8 +1,10 @@
 # 最新动态
 
 * 2020/06/18
-  * 完善mem allocator pool
   * 增加文档《内存分配》
+  * 完善mem allocator pool
+  * 增加函数check\_button\_get\_checked\_button
+  * 完善API注释（感谢俊杰提供补丁）。
 
 * 2020/06/17
   * 完善 edit（感谢培煌提供补丁）